Renamed manwhitney thumbnail aswell
