+++ src/components/charts/MannWhitneyUThumbnailView.tsx
